requirements.txt
fastapi
uvicorn
openai
python-dotenv
openai==0.28
